Finish spacing and padding page
The font in the legend tags seems to be different than the rest of the page, but trying to change that in the css didn't work, so I think that's causing the minor size disparity. Tried splitting the padding between both sides for the side by side ones, but they're still misaligned from the example, so not sure about that either.